Moving sampling tests to subfolder
